+++ src/Events/MfaMethodRemoved.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Ssntpl\Neev\Events;
+
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
+
+class MfaMethodRemoved
+{
+    use Dispatchable;
+    use SerializesModels;
+
+    public function __construct(
+        public $user,
+        public string $method,
+    ) {
+    }
+}