+++ src/Http/Controllers/Auth/OAuthController.php
@@ -3,6 +3,7 @@
 namespace Ssntpl\Neev\Http\Controllers\Auth;
 
 use Exception;
+use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
@@ -108,6 +109,9 @@ class OAuthController extends Controller
             return null;
         }
         DB::commit();
+
+        event(new Registered($user));
+
         return $user;
     }